requirements.txt
aiohttp
asyncpg
discord.py
python-dotenv
PyNaCl
pykakasi
